backend/pyproject.toml
[project]
name = "dining-hall-backend"
version = "0.1.0"
description = "Backend for dining hall recommendation app"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
